@@ -1,6 +1,6 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 import { AppRoutingModule } from './app-routing.module';
 import { AppComponent } from './app.component';
@@ -11,22 +11,15 @@ import { TableComponent } from './components/table/table.component';
 import { FooterComponent } from './components/footer/footer.component';
 import { NgOptimizedImage } from '@angular/common';
 
-@NgModule({
-  declarations: [
-    AppComponent,
-    HeaderComponent,
-    TableComponent,
-    FooterComponent
-  ],
-  imports: [
-    BrowserModule,
-    AppRoutingModule,
-    HttpClientModule,
-    BrowserAnimationsModule, 
-    MaterialModule,
-    NgOptimizedImage
-  ],
-  providers: [],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent,
+        HeaderComponent,
+        TableComponent,
+        FooterComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        AppRoutingModule,
+        BrowserAnimationsModule,
+        MaterialModule,
+        NgOptimizedImage], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class AppModule { }
